requirements.txt
pandas
numpy
pyyaml
matplotlib
seaborn
jupyter
scikit-learn
fastapi
uvicorn[standard]

